LayoutTests/html5lib/runner-expected-html5.txt

CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 0: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 0: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 0: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 2: PASS
CONSOLE MESSAGE: line 2: FOO<span>BAR</span>BAZ
Content-Type: text/plain
resources/tests1.dat: PASS

resources/tests2.dat: PASS

resources/tests3.dat: PASS

resources/tests4.dat:
3
5
6
7

Test 3 of 7 in resources/tests4.dat failed. Input:
textarea content with <em>pseudo</em> <foo>markup
Got:
| "textarea content with "
| <em>
|   "pseudo"
| " "
| <foo>
|   "markup"
Expected:
| "textarea content with <em>pseudo</em> <foo>markup"

Test 5 of 7 in resources/tests4.dat failed. Input:
</plaintext>
Got:
| 
Expected:
| "</plaintext>"

Test 6 of 7 in resources/tests4.dat failed. Input:
setting html's innerHTML
Got:
| "setting html's innerHTML"
Expected:
| <head>
| <body>
|   "setting html's innerHTML"

Test 7 of 7 in resources/tests4.dat failed. Input:
<title>setting head's innerHTML</title>
Got:
| 
Expected:
| <title>
|   "setting head's innerHTML"
resources/tests5.dat: PASS

resources/tests6.dat:
27
30
37
44
45
47

Test 27 of 51 in resources/tests6.dat failed. Input:
foo<col>
Got:
| 
Expected:
| <col>

Test 30 of 51 in resources/tests6.dat failed. Input:
</frameset><frame>
Got:
| 
Expected:
| <frame>

Test 37 of 51 in resources/tests6.dat failed. Input:
<caption><col><colgroup><tbody><tfoot><thead><tr>
Got:
| <caption>
| <colgroup>
| <tbody>
| <tfoot>
| <thead>
|   <tr>
Expected:
| <tr>

Test 44 of 51 in resources/tests6.dat failed. Input:
</table><tr>
Got:
| <tr>
Expected:
| <tbody>
|   <tr>

Test 45 of 51 in resources/tests6.dat failed. Input:
<body></body></html>
Got:
| 
Expected:
| <head>
| <body>

Test 47 of 51 in resources/tests6.dat failed. Input:
<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN"><html></html>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
Expected:
| <!DOCTYPE html "-//W3C//DTD HTML 4.01//EN" "">
| <html>
|   <head>
|   <body>
resources/tests7.dat:
24
30

Test 24 of 30 in resources/tests7.dat failed. Input:
<body>X</body></body>
Got:
| "X"
Expected:
| <head>
| <body>
|   "X"

Test 30 of 30 in resources/tests7.dat failed. Input:
<select><keygen>
Got:
| <html>
|   <head>
|   <body>
|     <select>
|     <keygen>
|       <option>
|         "2048 (High Grade)"
|       <option>
|         "1024 (Medium Grade)"
|       <option>
|         "512 (Low Grade)"
Expected:
| <html>
|   <head>
|   <body>
|     <select>
|     <keygen>
resources/tests8.dat: PASS

resources/tests9.dat: PASS

resources/tests10.dat: PASS

resources/tests11.dat: PASS

resources/tests12.dat: PASS

resources/tests14.dat: PASS

resources/tests15.dat: PASS

resources/tests16.dat: PASS

resources/tests17.dat: PASS

resources/tests18.dat: PASS

resources/tests19.dat:
1
22
24
63
77
78
80
81
101
102

Test 1 of 102 in resources/tests19.dat failed. Input:
<!doctype html><math><mn DefinitionUrl="foo">
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mn>
|         definitionurl="foo"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mn>
|         definitionURL="foo"

Test 22 of 102 in resources/tests19.dat failed. Input:
<!doctype html><h1><div><h3><span></h1>foo
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <h1>
|       <div>
|         <h3>
|           <span>
|     "foo"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <h1>
|       <div>
|         <h3>
|           <span>
|         "foo"

Test 24 of 102 in resources/tests19.dat failed. Input:
<!doctype html><h3><li>abc</h2>foo
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <h3>
|       <li>
|         "abcfoo"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <h3>
|       <li>
|         "abc"
|     "foo"

Test 63 of 102 in resources/tests19.dat failed. Input:
<!doctype html><keygen><frameset>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <keygen>
|       <option>
|         "2048 (High Grade)"
|       <option>
|         "1024 (Medium Grade)"
|       <option>
|         "512 (Low Grade)"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <keygen>

Test 77 of 102 in resources/tests19.dat failed. Input:
<html>�<frameset></frameset>
Got:
| <html>
|   <head>
|   <body>
|     "�"
Expected:
| <html>
|   <head>
|   <frameset>

Test 78 of 102 in resources/tests19.dat failed. Input:
<html> � <frameset></frameset>
Got:
| <html>
|   <head>
|   <body>
|     "� "
Expected:
| <html>
|   <head>
|   <frameset>

Test 80 of 102 in resources/tests19.dat failed. Input:
<html>��<frameset></frameset>
Got:
| <html>
|   <head>
|   <body>
|     "��"
Expected:
| <html>
|   <head>
|   <frameset>

Test 81 of 102 in resources/tests19.dat failed. Input:
<html>�
 <frameset></frameset>
Got:
| <html>
|   <head>
|   <body>
|     "�
 "
Expected:
| <html>
|   <head>
|   <frameset>

Test 101 of 102 in resources/tests19.dat failed. Input:
<!doctype html><figcaption><article></figcaption>a
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <figcaption>
|       <article>
|         "a"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <figcaption>
|       <article>
|     "a"

Test 102 of 102 in resources/tests19.dat failed. Input:
<!doctype html><summary><article></summary>a
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <summary>
|       <article>
|         "a"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <summary>
|       <article>
|     "a"
resources/tests20.dat:
16

Test 16 of 25 in resources/tests20.dat failed. Input:
<!doctype html><p><button><table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <p>
|         <table>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <table>
resources/webkit01.dat: PASS

resources/doctype01.dat:
14
15
16
23
24
25
26
27
29
30
32
33
34
35
36
37

Test 14 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato SYSTEM 'taco"'>Hello
Got:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato "" "taco"">
| <html>
|   <head>
|   <body>
|     "Hello"

Test 15 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato SYSTEM "taco">Hello
Got:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato "" "taco">
| <html>
|   <head>
|   <body>
|     "Hello"

Test 16 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato SYSTEM "tai'co">Hello
Got:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato "" "tai'co">
| <html>
|   <head>
|   <body>
|     "Hello"

Test 23 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato PUBLIC "go'of">Hello
Got:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato "go'of" "">
| <html>
|   <head>
|   <body>
|     "Hello"

Test 24 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato PUBLIC 'go'of'>Hello
Got:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato "go" "">
| <html>
|   <head>
|   <body>
|     "Hello"

Test 25 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato PUBLIC 'go:hh   of' >Hello
Got:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato "go:hh   of" "">
| <html>
|   <head>
|   <body>
|     "Hello"

Test 26 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato PUBLIC "W3C-//dfdf" SYSTEM ggg>Hello
Got:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato "W3C-//dfdf" "">
| <html>
|   <head>
|   <body>
|     "Hello"

Test 27 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01//EN"
   "http://www.w3.org/TR/html4/strict.dtd">Hello
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE html "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
| <html>
|   <head>
|   <body>
|     "Hello"

Test 29 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
"http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
Expected:
| <!DOCTYPE html "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
| <html>
|   <head>
|   <body>

Test 30 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Frameset//EN"
"http://www.w3.org/TR/xhtml1/DTD/xhtml1-frameset.dtd">
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
Expected:
| <!DOCTYPE html "-//W3C//DTD XHTML 1.0 Frameset//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-frameset.dtd">
| <html>
|   <head>
|   <body>

Test 32 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE html PUBLIC
  "-//WAPFORUM//DTD XHTML Mobile 1.0//EN"
    "http://www.wapforum.org/DTD/xhtml-mobile10.dtd">
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
Expected:
| <!DOCTYPE html "-//WAPFORUM//DTD XHTML Mobile 1.0//EN" "http://www.wapforum.org/DTD/xhtml-mobile10.dtd">
| <html>
|   <head>
|   <body>

Test 33 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE HTML SYSTEM "http://www.w3.org/DTD/HTML4-strict.dtd"><body><b>Mine!</b></body>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       "Mine!"
Expected:
| <!DOCTYPE html "" "http://www.w3.org/DTD/HTML4-strict.dtd">
| <html>
|   <head>
|   <body>
|     <b>
|       "Mine!"

Test 34 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01//EN""http://www.w3.org/TR/html4/strict.dtd">
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
Expected:
| <!DOCTYPE html "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
| <html>
|   <head>
|   <body>

Test 35 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01//EN"'http://www.w3.org/TR/html4/strict.dtd'>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
Expected:
| <!DOCTYPE html "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
| <html>
|   <head>
|   <body>

Test 36 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE HTML PUBLIC"-//W3C//DTD HTML 4.01//EN"'http://www.w3.org/TR/html4/strict.dtd'>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
Expected:
| <!DOCTYPE html "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
| <html>
|   <head>
|   <body>

Test 37 of 37 in resources/doctype01.dat failed. Input:
<!DOCTYPE HTML PUBLIC'-//W3C//DTD HTML 4.01//EN''http://www.w3.org/TR/html4/strict.dtd'>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
Expected:
| <!DOCTYPE html "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
| <html>
|   <head>
|   <body>
resources/scriptdata01.dat: PASS

resources/html5test-com.dat:
7
9
10
11

Test 7 of 24 in resources/html5test-com.dat failed. Input:
&lang;&rang;
Got:
| <html>
|   <head>
|   <body>
|     "〈〉"
Expected:
| <html>
|   <head>
|   <body>
|     "⟨⟩"

Test 9 of 24 in resources/html5test-com.dat failed. Input:
&ImaginaryI;
Got:
| <html>
|   <head>
|   <body>
|     "&ImaginaryI;"
Expected:
| <html>
|   <head>
|   <body>
|     "ⅈ"

Test 10 of 24 in resources/html5test-com.dat failed. Input:
&Kopf;
Got:
| <html>
|   <head>
|   <body>
|     "&Kopf;"
Expected:
| <html>
|   <head>
|   <body>
|     "𝕂"

Test 11 of 24 in resources/html5test-com.dat failed. Input:
&notinva;
Got:
| <html>
|   <head>
|   <body>
|     "&notinva;"
Expected:
| <html>
|   <head>
|   <body>
|     "∉"
resources/entities01.dat:
2
5

Test 2 of 68 in resources/entities01.dat failed. Input:
FOO&gtBAR
Got:
| <html>
|   <head>
|   <body>
|     "FOO&gtBAR"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO>BAR"

Test 5 of 68 in resources/entities01.dat failed. Input:
I'm &notit; I tell you
Got:
| <html>
|   <head>
|   <body>
|     "I'm &notit; I tell you"
Expected:
| <html>
|   <head>
|   <body>
|     "I'm ¬it; I tell you"
resources/entities02.dat: PASS

resources/comments01.dat: PASS

resources/adoption01.dat:
12

Test 12 of 12 in resources/adoption01.dat failed. Input:
<table>A<td>B</td>C</table>
Got:
| <html>
|   <head>
|   <body>
|     "AC"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "B"
Expected:
| <html>
|   <head>
|   <body>
|     "A"
|     "C"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "B"
resources/adoption02.dat:
2

Test 2 of 2 in resources/adoption02.dat failed. Input:
<b id="b">1<i id="i">2<p>3<script>document.getElementById("b").id = "c";document.getElementById("i").id = "j"</script></b>4
Got:
| <html>
|   <head>
|   <body>
|     <b>
|       id="c"
|       "1"
|       <i>
|         id="j"
|         "2"
|     <i>
|       id="j"
|       <p>
|         <b>
|           id="c"
|           "3"
|           <script>
|             "document.getElementById("b").id = "c";document.getElementById("i").id = "j""
|         "4"
Expected:
| <html>
|   <head>
|   <body>
|     <b>
|       id="c"
|       "1"
|       <i>
|         id="j"
|         "2"
|     <i>
|       id="i"
|       <p>
|         <b>
|           id="b"
|           "3"
|           <script>
|             "document.getElementById("b").id = "c";document.getElementById("i").id = "j""
|         "4"
resources/inbody01.dat: PASS

resources/isindex.dat: PASS

resources/tables01.dat: PASS

resources/tricky01.dat: PASS

resources/scripted/adoption01.dat:
1

Test 1 of 1 in resources/scripted/adoption01.dat failed. Input:
<p><b id="A"><script>document.getElementById("A").id = "B"</script></p>TEXT</b>
Got:
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|         id="B"
|         <script>
|           "document.getElementById("A").id = "B""
|     <b>
|       id="B"
|       "TEXT"
Expected:
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|         id="B"
|         <script>
|           "document.getElementById("A").id = "B""
|     <b>
|       id="A"
|       "TEXT"
resources/scripted/webkit01.dat: PASS
#EOF
